Split runtime voice seam into settings/executor/coordinator modules
