+++ src/yoyopod/runtime/voice/__init__.py
@@ -1,0 +1,14 @@
+"""Runtime-owned voice orchestration split by responsibility."""
+
+from __future__ import annotations
+
+from .coordinator import VoiceRuntimeCoordinator
+from .executor import VoiceCommandExecutor
+from .settings import VoiceCommandOutcome, VoiceSettingsResolver
+
+__all__ = [
+    "VoiceCommandExecutor",
+    "VoiceCommandOutcome",
+    "VoiceRuntimeCoordinator",
+    "VoiceSettingsResolver",
+]
